requirements.txt
openai
pinecone-client
langchain
supabase
fastapi
python-dotenv
sqlalchemy
requests
httpx
uvicorn
pydantic
python-multipart
